[SERVICE ADDITION] there is now a service called 'exam-quest-picture' that aims to picture all the questions from the students AND the templates. Now, questionTypeInteraction will be used to recuperate the type of a question and to call the exam-quest-picture service for then sending all the captures of the question to the appropriate API (if there is one)'
