+++ src/main/webapp/app/entities/exam-sheet/service/exam-quest-picture-service.service.spec.ts
@@ -1,0 +1,16 @@
+import { TestBed } from '@angular/core/testing';
+
+import { ExamQuestPictureServiceService } from './exam-quest-picture-service.service';
+
+describe('ExamQuestPictureServiceService', () => {
+  let service: ExamQuestPictureServiceService;
+
+  beforeEach(() => {
+    TestBed.configureTestingModule({});
+    service = TestBed.inject(ExamQuestPictureServiceService);
+  });
+
+  it('should be created', () => {
+    expect(service).toBeTruthy();
+  });
+});
